ml-service/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
python-jose[cryptography]
numpy
torch / torchvision
opencv-python
scikit-learn 